Add an initial configuration for ESLint
This adds ESLint as a dependency for the project. Though it may be
more efficient to install it globally, it's kind of nice to let it
be known explicitly what is required for a project.

It also adds an NPM script - using `npm run lint:es` will lint any
`.js` files in the `src` directory of any directory within the repo.

In addition to the above, I've also installed the `ESLint` extension
for VSCode (https://github.com/Microsoft/vscode-eslint).

Lastly, this adds a rudimentary configuration for ESLint. Globals
from the `browser`, `node`, `amd`, and `es6` are enabled, and the
base `eslint:recommended` rules are enabled. We will probably want
to explore setting up a `.eslintrc` file to add a more opinionated
configuration for our projects.