Add an intial configuration for stylelint
I'm not certain this is the tool we want to use for CSS linting, so
I didn't go too far into it.

This adds the `stylelint` and `stylelint-config-recommended` packages
as dependencies, and an NPM script - `npm run lint:css` - for linting
all CSS files in the `src` directory of any directory within the repo.

So far it's just using the recommended sharable config for stylelint,
so we'll no doubt want to configure things further to meet our needs.

There is a VSCode extension that's worth exploring. It seems like
VSCode already has some integrated CSS linting in place - if that is
indeed the case, we may want to look into disabling it.