Add an initial configuration for Babel
This adds a few Babel-related dependecies for setting up a basic
configuration for compiling any `.js` files in the `src` directory
of any directory within the repo.

Using `npm run build` will output the compiled files in the directory's
`dist` directory. Using `npm start` does the same thing and is supposed
to watch for changes, but doesn't seem to be working...

This uses the `@babel/preset-env` so we can use the latest JavaScript.
I also opted to remove comments and minify the compiled files.

This also adds a `browserslist` property, which Babel and other tools
(stylelint and autoprefixer, to name two) use to apply polyfills and
fallbacks.