Refactor scroll event handling
This adds a listener for the `scroll` event and replaces the
`EventListenerOptions` object with the boolean `true` for the last
parameter. Apparently the scroll event is passive by default. I'll
admit that I don't fully grasp all of that.

This also moves the scroll event optimization into an intermediary
function to clean things up a bit, reorganizes methos, and adds
spacing throughout for consistency.