Update toggle handling
This adds a button instead of hijacking the parent item's anchor
as a section toggle.

It also adds some rudimentary styles. It might be fun to `use` a
very simple SVG for the button icon. Replacing the default `focus`
styles with a `box-shadow` or something that can leverage
`border-radius` might be nice, too, if we keep the rotate animation.

This should probably only apply to the vertical navigation - the
horizontal will likely be handled differently.